+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/dashboard/link/DashboardLinkType.kt
@@ -1,0 +1,16 @@
+package ru.yandex.money.tools.grafana.dsl.dashboard.link
+
+/**
+ * Type of dashboard's link
+ */
+enum class DashboardLinkType(val value: String) {
+    /**
+     * Link to other dashboards
+     */
+    DASHBOARDS("dashboards"),
+
+    /**
+     * Link to arbitrary url address
+     */
+    URI_LINK("link")
+}